--- ProyectoWipStudios/Assets/Prefabs/PrototypeEnemy.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 53d438789353e4d469ced5e722bc5575
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
