--- ProyectoWipStudios/Assets/Scripts/AI/ProtoAI.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 19e9cf619abbc6f4aa6f1a955b7e1bb9
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
